avoid function duplication
